[Onboarding]: NV-3415 | Feature flag and conditional rendering setup (#5169)
- Add feature flag
- Refactor app entry point (Routes, App, Providers, initialization)
