+++ apps/web/src/config/index.ts
@@ -16,6 +16,7 @@ import {
   IS_TEMPLATE_STORE_ENABLED,
   IS_MULTI_TENANCY_ENABLED,
   IS_TRANSLATION_MANAGER_ENABLED,
+  IS_IMPROVED_ONBOARDING_ENABLED,
 } from '@novu/shared-web';
 
 export {
@@ -36,6 +37,7 @@ export {
   IS_TEMPLATE_STORE_ENABLED,
   IS_MULTI_TENANCY_ENABLED,
   IS_TRANSLATION_MANAGER_ENABLED,
+  IS_IMPROVED_ONBOARDING_ENABLED,
 };
 
 export const IS_EU_ENV = (ENV === 'production' || ENV === 'prod') && API_ROOT.includes('eu.api.novu.co');
