+++ apps/web/src/pages/get-started/GetStartedPage.tsx
@@ -5,15 +5,15 @@ import PageContainer from '../../components/layout/components/PageContainer';
 import PageHeader from '../../components/layout/components/PageHeader';
 import { GetStartedTabs } from './components/get-started-tabs/GetStartedTabs';
 
+const PAGE_TITLE = 'Get started';
+
 export function GetStartedPage() {
   return (
-    <>
-      <PageContainer title="Get Started">
-        <HeaderLayout>
-          <PageHeader title="Get Started" />
-        </HeaderLayout>
-        <GetStartedTabs />
-      </PageContainer>
-    </>
+    <PageContainer title={PAGE_TITLE}>
+      <HeaderLayout>
+        <PageHeader title={PAGE_TITLE} />
+      </HeaderLayout>
+      <GetStartedTabs />
+    </PageContainer>
   );
 }
