+++ libs/shared/src/types/feature-flags/feature-flags.ts
@@ -7,4 +7,5 @@ export enum FeatureFlagsKeysEnum {
   IS_API_IDEMPOTENCY_ENABLED = 'IS_API_IDEMPOTENCY_ENABLED',
   IS_TRANSLATION_MANAGER_ENABLED = 'IS_TRANSLATION_MANAGER_ENABLED',
   IS_API_EXECUTION_LOG_QUEUE_ENABLED = 'IS_API_EXECUTION_LOG_QUEUE_ENABLED',
+  IS_IMPROVED_ONBOARDING_ENABLED = 'IS_IMPROVED_ONBOARDING_ENABLED',
 }
